requirements.txt
aiogram==3.13.1
fastapi==0.115.2
python-dotenv==1.0.1
uvicorn

httpx~=0.28.1